+++ lib/i18n/strings.g.dart
@@ -6,7 +6,7 @@
 /// Locales: 1
 /// Strings: 768
 ///
-/// Built on 2026-06-03 at 05:50 UTC
+/// Built on 2026-06-08 at 19:34 UTC
 
 // coverage:ignore-file
 // ignore_for_file: type=lint, unused_import